requirements.txt
rich~=13.5.2
click~=8.1.6
prompt_toolkit~=3.0.39
pyperclip~=1.8.2
pyyaml~=6.0.1
../mypass-requests
../mypass-logman
../mypass-tokens

Pygments~=2.15.1
setuptools~=65.5.1
